Move accounts template to credentials dir.
